requirements.txt
appnope==0.1.4
APScheduler==3.10.4
asttokens==3.0.0
attrs==24.3.0
blinker==1.9.0
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
colorhash==2.0.0
comm==0.2.2
configparser==7.1.0
contourpy==1.3.1
cycler==0.12.1
debugpy==1.8.11
decorator==5.1.1
exceptiongroup==1.2.2
executing==2.1.0
Flask==3.1.0
Flask-Cors==5.0.0
Flask-MonitoringDashboard==3.3.2
fonttools==4.55.3
idna==3.10
imbalanced-learn==0.13.0
imblearn==0.0
importlib_metadata==8.5.0
ipykernel==6.29.5
ipython==8.31.0
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.5
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.8
kneed==0.8.5
MarkupSafe==3.0.2
matplotlib==3.10.0
matplotlib-inline==0.1.7
more-itertools==10.5.0
nest-asyncio==1.6.0
numpy==2.2.1
packaging==24.2
pandas==2.2.3
parso==0.8.4
pexpect==4.9.0
pillow==11.1.0
platformdirs==4.3.6
prompt_toolkit==3.0.48
psutil==6.1.1
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.19.1
pyparsing==3.2.1
pyrsistent==0.20.0
python-dateutil==2.9.0.post0
pytz==2024.2
pyzmq==26.2.0
referencing==0.35.1
regexp==0.1
requests==2.31.0
rpds-py==0.22.3
scikit-learn==1.6.1
scipy==1.15.1
seaborn==0.13.2
six==1.17.0
sklearn-compat==0.1.3
SQLAlchemy==2.0.37
stack-data==0.6.3
threadpoolctl==3.5.0
tornado==6.4.2
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2024.2
tzlocal==2.0.0
urllib3==2.3.0
wcwidth==0.2.13
Werkzeug==3.1.3
xgboost==2.1.3
zipp==3.21.0
